requirements.txt
fastapi
uvicorn
requests
beautifulsoup4
pydantic
google-generativeai
python-dotenv
duckduckgo-search
